LayoutTests/imported/w3c/web-platform-tests/mathml/presentation-markup/operators/operator-dictionary-symmetric-006-expected.txt


FAIL Operator dictionary chunk 1 - symmetric assert_approx_equals: Symmetric property for ‖ postfix should be 'true' expected 150 +/- 1 but got 100
PASS Operator dictionary chunk 2 - symmetric
PASS Operator dictionary chunk 3 - symmetric
PASS Operator dictionary chunk 4 - symmetric
symmetric for "‖" (postfix):
‖
 VS
‖
